+++ NEWS.md
@@ -1,6 +1,18 @@
+# npfda 0.1-4 (2023-05-15) 
+
+* Added S3 generic `npf.fit()` (with methods for `npf.data` 
+  and `npf.locpol` classes).
+  
+
 # npfda 0.1-3 (2023-05-12) 
 
 * Added `h.cv()` S3 methods for `npf.bin` and `npf.bin.res2` classes.
+
+* Added `npf.model()` (constructor of `npf.model`-class) 
+
+* Added `residuals()` S3 methods for `npf.locpol` and `npf.model` classes.
+
+* Added `plot()` S3 methods for `npf.locpol` and `npf.model` classes.
   
 
 # npfda 0.1-2 (2023-05-10) 
@@ -14,11 +26,11 @@
 
 # npfda 0.1-1 (2023-05-07) 
 
-* Added `npf.bin.res2()` and S3 generic `np.var()` (with methods for classes 
-  `npf.locpol` and `npf.bin.res2`; results `data` components are not fully compatible 
+* Added `npf.bin.res2()` and S3 generic `np.var()` (with methods for `npf.locpol` 
+  and `npf.bin.res2` classes; results `data` components are not fully compatible 
   with the `npsp` package). 
   
-* Added `predict()` S3 methods for classes `npf.locpol` and `npf.var`.
+* Added `predict()` S3 methods for `npf.locpol` and `npf.var` classes.
 
 * Changes in the `data` component of the return value of `npf.binning()`
   (results are now not fully compatible with the `npsp` package;